+++ app/nux/starter-content.php
@@ -10,6 +10,7 @@
  */
 
 // An annoying hack to ensure we can always importer starter content.
+// @codingStandardsIgnoreStart
 if ( isset( $_GET['starter-content-redirect'] ) ) {
 	update_option( 'fresh_site', 1 );
 
@@ -26,6 +27,7 @@ if ( isset( $_GET['starter-content-redirect'] ) ) {
 
 	exit();
 }
+// @codingStandardsIgnoreEnd
 
 /**
  * Starter content.